wip echo, added drawLineAA

+++ include/compose/modifiers/Rect.h
@@ -5,8 +5,8 @@ namespace Compose
 {
   struct Point
   {
-    int x = {};
-    int y = {};
+    float x = {};
+    float y = {};
   };
 
   struct Rect
